requirements.txt
black
ruff
isort
pydantic
python-dotenv
facebook-business
google-cloud-storage==2.14.0
google-cloud-videointelligence==2.13.0
google-generativeai==0.5.4
gspread==5.12.4
oauth2client==4.1.3
pandas
selenium
selenium-wire
webdriver-manager
requests
beautifulsoup4
pydantic-settings 